api-backend/backend/requirements.txt
fastapi
uvicorn
python-dotenv
asyncio
httpx
google-genai
agno
openai
pypdf
duckduckgo-search
ddgs
lancedb
tantivy
pylance
pandas
rich
markdown
aiofiles
beautifulsoup4
